requirements.txt
2captcha-python==1.5.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
better-automation==2.0.2
better-proxy==1.2.0
bitarray==3.0.0
blessed==1.20.0
capmonstercloudclient==1.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==1.0.2
curl_cffi==0.6.2
cytoolz==1.0.0
editor==1.6.6
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
fake-useragent==1.5.1
Faker==33.0.0
frozenlist==1.5.0
hexbytes==0.3.1
idna==3.10
inquirer==3.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.2
lru-dict==1.2.0
lxml==5.3.0
multidict==6.1.0
parsimonious==0.10.0
propcache==0.2.0
protobuf==5.28.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.1.1
pydantic_core==2.4.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyunormalize==16.0.0
pyuseragents==1.0.5
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
rpds-py==0.21.0
runs==1.2.2
six==1.16.0
soupsieve==2.6
tenacity==8.5.0
termcolor==2.5.0
tinydb==4.8.2
toolz==1.0.0
tqdm==4.67.0
git+https://github.com/Injustive/tweepy-self#egg=tweepy-self
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
web3==6.15.1
websockets==14.1
xmod==1.8.1
yarl==1.18.0
openpyxl==3.1.5
SQLAlchemy==2.0.38
aiosqlite==0.20.0
opencv-python==4.11.0.86
numpy==2.2.3
patchright==1.50.0
cryptography==44.0.2
pyfiglet==1.0.2
colorama==0.4.6
primp==0.14.0
PyYAML==6.0.2
aiofiles==24.1.0
prettytable==3.15.1
openai
playwright
